:recycle: Refactor balanceChart distributing date by series
